requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
qdrant-client==1.5.4
aiohttp==3.8.5
websockets==11.0.3
numpy==1.25.2
openai==1.3.5
python-multipart==0.0.6 